Add Smudge spotify controller
It does work but I haven't been able to automatically enable the keyboard menu
shortcuts. I think it may be due to some kind of issue with the package itself?
All the recommended methods fallover as the smudge stuff has not been defined
yet. Event `(require 'smudge)` fails.

I've left the command commented out in the init file though, so the sexp can be
run if needed.