Fix email deletion on Gmail accounts
The problem previously was that whenever the "trashed" flag was set on an email
Gmail would stick it into "All Mail". With this update we are overriding the
"trash" action by removing the `+T` flag change.

Now the mail will be moved to the Gmail trash folder, and will be deleted within
30 days just as it should. This works in both directions.