Style posframe borders as yellow
For some reason which I have not yet worked out the styling for posframes gets
lost at some point. When this happens it falls back to the normal background
colour of the main editor (which isn't the end of the world, but not as nice),
and the default internal frame colour.

By setting all the borders to yellow we will always at least get a border to
separate from what is underneath no matter which type of styling is happening.