Fix loading of font on initial load
Until now I have always had to reload the init file on load for Cascadia to
render. This seems to have fixed it, in a bit of a hacky way!

1. Tell Spacemacs to set the detault font to `nil`.
   I am not sure why it fails when attempting Cascadia, but this is actually the
   last thing to be run, and even when the correct font was being rendered during
   startup that final command would reset it to Menlo
2. Load the theme file twice.
   I have absolutely no idea why this is necessary, but it is.

As part of this all the font stuff has also been extracted to its own file. I'm
planning to tidy up and separate out the whole init config at some point anyway
so this is a start.