Use Monaspace fonts
This font family has multiple faces all with the same width, so we can use a
script type font for comments in the same file as a standard coding face.

I would like to use the "Var" versions of the face, which uses texture healing
to make it render more like a variable width font, however only some of these
are available to Emacs for some reason, and that does not include the exact
version I need, resulting in all code being rendered in italics.