Use org-roam to create daily task lists
Can create new todo items for each day. I think the way this will work is to
have a backlog in the Tasks.org file, and each day either create new items to
do, or refile them from the backlog or previous days.

Going forward we can create shortcuts for doing this, and also scheduling
automatically so that it shows up in the agenda.