Fix starring of Gmail emails
Instead of adding a "Flagged" flag, Gmail puts them in the "Starred" folder.
This is not one of the default supported folders in mu4e (such as "Send" or
"Trashed") so we have to do a bit of string manipulation to find the correct
folder in the context of the given message.