Remove helm-posframe
It doesn't work well at all with the search. Fine for simple tasks, but if you
wante to, for example, preview the matched item without opening it, then you
couldn't see it and it would often just crash.

Keeping it for transients and any-key though, still nice to have those in the
middle of the screen.