Do not recompile webpack
The reason for having this flag in the first place is to prevent it happening
multiple times in different threads. Locally I almost never parallelise running
the test suite so there is no need.

It doesn't work currently when running tests from within Emacs, I assume I need
to do something to reload the env vars into it or something.