Be more explicit about ruby lsp
Previously this did work as default, but sometimes the default would change, so
better to be explicit.

We definitely want the LSP for symbol navigation, however it's formatter
requires prettier, which is too opinionated for a shared project like Cleo
unless everyone is onboard (unlikely to happen). So have added a shortcut
`SPC o f` to call it.