Restore ChatGPT
This is now a supported layer in Spacemacs, called llm-client but it uses the
same gptel package.

There is a bug which prevents it opening a new buffer, so there is a bit of
hackery that has to happen. Passing a string to the command creates a new buffer
with that name.