Add treesitter mode, and typespec grammar
I had to recompile Emacs with the `--with-tree-sitter` option, but then
we do also have to add the grammar. There are plenty of languages in the
`tree-sitter-langs` packages, but not typespec. This adds the location
but it will still need to be manually added with the command listed in
the org file comment.