Add email back... again
I just can't deal with email in Gmail so I have to make this work somehow.

The key issue seems to be that while emails can have multiple labels in Gmail
they can't be in multiple folders in mu. mu is happy to sync duplicate versions
of the same email, if that's what Gmail's imap is giving it in the case of mail
with multiple labels.

One issue with this is with sent mail. From mu's side it is not possible to
archive a thread and keep it in sent mail too. However the solution appears to
be to refile only the incoming emails, leaving the ones from me alone. This
works well as with some of the mails being archived, and some in sent, the full
thread is diplayed on both.

I'm also going to stop syncing the "All Mail" folder completely, and refile into
"_Archive". Every mail in every folder would be duplicated into "All Mail" which
is problematic for mu's single folder model, but also doubles the number of
mails to be downloaded.

Fingers crossed!!