Restore key monitor to close hyprswitcher
I have still found no way to handle a Super release using Hyprland keybindings
here. Once it has been used as a modifier it seems all bets are off.

I tried starting the script when the switcher is launched, but funny things go
on when it is spawned from hyprland and, again, the release appears to be
swallowed.

This works as it did before, but using evtest rather than the previous method
which was a debug command, this should be much less CPU intensive.