Update prompt from scrape to round edges
Annoyingly it's impossible to set a foreground color to be the background, and
it's impossible to explicitly set a bg color, it depends on transparency and
things like that. The scrape is a foreground only character so it doesn't work
on the left.

The round probably fits better with the Omarchy style anyway.